+++ directional_light.h
@@ -10,7 +10,7 @@ class DirectionalLight: public Light {
   	Ray* get_shadow_ray(ThreeDVector* postion);
   	vector<Ray*> get_shadow_rays(ThreeDVector* position, int sample_size);
   	void apply_transformation(Eigen::Matrix4f transformation);
-  	DirectionalLight(float, float, float, float, float, float);
+  	DirectionalLight(long double, long double, long double, long double, long double, long double);
   	~DirectionalLight();
 };
